pyproject.toml
[build-system]
requires = [
    "setuptools>=64",
    "wheel",
    "torch==1.13.1",
    "numpy>=1.24.4,<2.0",
    "pybind11",
]
build-backend = "setuptools.build_meta"

[project]
name = "bevnext-sam2"
version = "1.0.0"
description = "BEVNeXt + SAM2: Unified 3D Object Detection and Segmentation"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "BEVNeXt-SAM2 Contributors"},
]
keywords = ["computer vision", "3D object detection", "segmentation", "BEV", "SAM"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8.0"
dependencies = [
    "torch==1.13.1",
    "torchvision==0.14.1",
    "numpy>=1.24.4,<2.0",
    "tqdm>=4.66.1",
    "hydra-core>=1.3.2",
    "iopath>=0.1.10",
    "pillow>=9.4.0",
    "networkx>=2.2,<2.3",
    "numba>=0.55.0",
    "plyfile",
    "scikit-image",
    "tensorboard",
    "trimesh>=2.35.39,<2.35.40",
]

[project.optional-dependencies]
all = [
    "matplotlib>=3.9.1",
    "jupyter>=1.0.0",
    "opencv-python>=4.7.0",
    "eva-decord>=0.6.1",
    "Flask>=3.0.3",
    "Flask-Cors>=5.0.0",
    "av>=13.0.0",
]
dev = [
    "black==24.2.0",
    "usort==1.0.2",
    "ufmt==2.0.0b2",
    "fvcore>=0.1.5.post20221221",
    "pandas>=2.2.2",
    "pycocotools>=2.0.8",
]

[project.urls]
Homepage = "https://github.com/your-repo/bevnext-sam2"

[tool.setuptools]
packages = ["bevnext", "sam2_module", "configs", "tools", "scripts"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.txt", "*.md"] 